+++ src/main/java/com/romoshi/bot/telegram/constant/BotStringConstant.java
@@ -2,13 +2,14 @@ package com.romoshi.bot.telegram.constant;
 
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @NoArgsConstructor
 public class BotStringConstant {
 
     public static final String START_STRING = "Привет, друг \uD83D\uDC4B \nЧто тебя интересует?";
     public static final String ADD_STRING = "Товар добавлен!";
     public static final String SALE_STRING = "Товар куплен!";
-
     public static final String PRODUCT_LIST = "Список товаров:";
     public static final String HAVE_NOT_PRODUCTS = "В магазине пока нет товаров.";
     public static final String SITE_STRING = "https://doctor-dyorina.tilda.ws/";
@@ -19,7 +20,7 @@ public class BotStringConstant {
     public static final String DEFAULT_PRODUCT_FILE_ID = "null";
     public static final String DEFAULT_PRODUCT_NAME = "Название товара";
     public static final String DEFAULT_DESCRIPTION = "Описание товара.";
-    public static final int DEFAULT_PRODUCT_PRICE = 100;
+    public static final BigDecimal DEFAULT_PRODUCT_PRICE = BigDecimal.valueOf(100);
 
     //Update, delete button
     public static final String UPDATE_MESSAGE = "Что хотите изменить?";
